Begin architecture buider and port of swin transformer
